+++ src/openhuman/agent/harness/subagent_runner/ops.rs
@@ -1,144 +1,37 @@
-//! Sub-agent execution runner.
+//! Sub-agent execution entry points and the inner tool-call loop.
 //!
-//! Given an [`AgentDefinition`] and a task prompt, this module:
+//! The public runner lives in [`run_subagent`]. It dispatches to
+//! [`run_typed_mode`] (narrow prompt + filtered tools) or
+//! [`run_fork_mode`] (prefix-replay) depending on the
+//! [`super::types::SubagentMode`] implied by the
+//! [`crate::openhuman::agent::harness::definition::AgentDefinition`].
 //!
-//! 1. Reads the [`ParentExecutionContext`] task-local set by the parent
-//!    [`crate::openhuman::agent::Agent::turn`].
-//! 2. Resolves the sub-agent's model name (inherit / hint / exact).
-//! 3. Filters the parent's tool registry per `definition.tools`,
-//!    `disallowed_tools`, and `skill_filter` (or, in `fork` mode,
-//!    inherits the parent's tools verbatim).
-//! 4. Builds a narrow system prompt that strips the sections the
-//!    definition asks to omit (`omit_identity`, `omit_memory_context`,
-//!    `omit_safety_preamble`, `omit_skills_catalog`).
-//! 5. Runs a slim inner tool-call loop using the parent's
-//!    [`crate::openhuman::providers::Provider`] and returns a single
-//!    text result. The intra-sub-agent history never leaks back to the
-//!    parent — the parent only sees one compact tool result.
-//!
-//! Token-saving levers in this runner:
-//! - **Narrow prompt** — typed sub-agents skip identity/memory/skills.
-//! - **Filtered tools** — fewer schemas in the request body.
-//! - **Cheaper model** — archetype hint usually selects a smaller model.
-//! - **Lower max iterations** — definition-controlled per archetype.
-//! - **No memory recall** — sub-agents skip per-turn memory loading entirely;
-//!   the parent has already injected the relevant context.
-//! - **Structural compaction** — sub-agent's tool-call history collapses
-//!   into a single tool result block in the parent's history.
-//! - **Fork-mode prefix replay** — `uses_fork_context` definitions
-//!   replay the parent's exact bytes for backend prefix-cache hits.
+//! Both modes delegate to [`run_inner_loop`] which drives provider
+//! calls and tool execution until the model returns without further
+//! tool calls (or the iteration budget is exhausted).
 
-use super::definition::{AgentDefinition, PromptSource, ToolScope};
-use super::fork_context::{current_fork, current_parent, ForkContext, ParentExecutionContext};
-use super::session::transcript;
+use std::collections::HashSet;
+use std::sync::Arc;
+use std::time::Instant;
+
+use super::extract_tool::ExtractFromResultTool;
+use super::handoff::{
+    build_handoff_placeholder, clean_tool_output, ResultHandoffCache,
+    HANDOFF_OVERSIZE_THRESHOLD_TOKENS,
+};
+use super::tool_prep::{
+    build_text_mode_tool_instructions, filter_tool_indices, is_welcome_only_tool,
+    load_prompt_source, top_k_for_toolkit,
+};
+use super::types::{SubagentMode, SubagentRunError, SubagentRunOptions, SubagentRunOutcome};
+use crate::openhuman::agent::harness::definition::{AgentDefinition, PromptSource};
+use super::super::fork_context::{current_fork, current_parent, ForkContext, ParentExecutionContext};
+use super::super::session::transcript;
 use crate::openhuman::context::prompt::{
     render_subagent_system_prompt, PromptContext, PromptTool, SubagentRenderOptions,
 };
 use crate::openhuman::providers::{ChatMessage, ChatRequest, Provider, ToolCall};
-use crate::openhuman::tools::{Tool, ToolCategory, ToolResult, ToolSpec};
-use async_trait::async_trait;
-use futures::stream::StreamExt;
-use regex::Regex;
-use std::collections::{HashMap, HashSet};
-use std::sync::{Arc, LazyLock, Mutex as StdMutex};
-use std::time::{Duration, Instant};
-use thiserror::Error;
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Public API
-// ─────────────────────────────────────────────────────────────────────────────
-
-/// Per-spawn options that override or augment what the
-/// [`AgentDefinition`] specifies. Built by `SpawnSubagentTool::execute`
-/// from the parent model's call arguments.
-#[derive(Debug, Clone, Default)]
-pub struct SubagentRunOptions {
-    /// Optional skill-id override (e.g. `"notion"`). When set, the
-    /// resolved tool list is further restricted to tools whose name
-    /// starts with `{skill}__`. Overrides `definition.skill_filter`.
-    pub skill_filter_override: Option<String>,
-
-    /// Optional Composio toolkit scope (e.g. `"gmail"`, `"notion"`).
-    /// When set, skill-category tools are further restricted to those
-    /// whose name starts with the uppercased `{toolkit}_` prefix, and
-    /// the sub-agent's rendered `Connected Integrations` section is
-    /// narrowed to only that toolkit's entry. Used by main/orchestrator
-    /// when spawning `integrations_agent` for a specific platform so the
-    /// sub-agent only sees one integration's tool catalogue.
-    pub toolkit_override: Option<String>,
-
-    /// Optional context blob the parent wants to inject before the
-    /// task prompt. Rendered as a `[Context]\n…\n` prefix.
-    pub context: Option<String>,
-
-    /// Stable id for tracing / DomainEvents (defaults to a UUID).
-    pub task_id: Option<String>,
-}
-
-/// Outcome of a single sub-agent run, returned to the parent.
-#[derive(Debug, Clone)]
-pub struct SubagentRunOutcome {
-    /// Unique identifier for this sub-task run.
-    pub task_id: String,
-    /// The ID of the agent archetype used (e.g., `researcher`).
-    pub agent_id: String,
-    /// The final text response produced by the sub-agent.
-    pub output: String,
-    /// How many LLM round-trips were performed during the run.
-    pub iterations: usize,
-    /// Total wall-clock duration of the run.
-    pub elapsed: Duration,
-    /// Which execution mode was used (Typed vs. Fork).
-    pub mode: SubagentMode,
-}
-
-/// Which prompt-construction path the runner took for a sub-agent.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SubagentMode {
-    /// Built a narrow, archetype-specific prompt with filtered tools.
-    Typed,
-    /// Replayed the parent's exact rendered prompt and history prefix.
-    Fork,
-}
-
-impl SubagentMode {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::Typed => "typed",
-            Self::Fork => "fork",
-        }
-    }
-}
-
-/// Errors the runner can surface to the parent. The parent receives a
-/// stringified version inside a tool result block.
-#[derive(Debug, Error)]
-pub enum SubagentRunError {
-    #[error("spawn_subagent called outside of an agent turn — no parent context available")]
-    NoParentContext,
-
-    #[error(
-        "fork-mode sub-agent requested but no ForkContext is set on the task-local. \
-         Did the parent agent forget to call `Agent::turn` with fork support?"
-    )]
-    NoForkContext,
-
-    #[error("agent definition '{0}' not found in registry")]
-    DefinitionNotFound(String),
-
-    #[error("failed to load archetype prompt from '{path}': {source}")]
-    PromptLoad {
-        path: String,
-        #[source]
-        source: std::io::Error,
-    },
-
-    #[error("provider call failed: {0}")]
-    Provider(#[from] anyhow::Error),
-
-    #[error("sub-agent exceeded maximum iterations ({0})")]
-    MaxIterationsExceeded(usize),
-}
+use crate::openhuman::tools::{Tool, ToolCategory, ToolSpec};
 
 /// Run a sub-agent based on its definition and a task prompt.
 ///
@@ -373,13 +266,13 @@ async fn run_typed_mode(
                 // too-narrow filter is worse than none — it starves the
                 // sub-agent and forces it to guess.
                 let top_k = top_k_for_toolkit(tk);
-                let filter_hits = super::tool_filter::filter_actions_by_prompt(
+                let filter_hits = super::super::tool_filter::filter_actions_by_prompt(
                     task_prompt,
                     &integration.tools,
                     top_k,
                 );
                 let selected: Vec<&crate::openhuman::context::prompt::ConnectedIntegrationTool> =
-                    if filter_hits.len() >= super::tool_filter::MIN_CONFIDENT_HITS {
+                    if filter_hits.len() >= super::super::tool_filter::MIN_CONFIDENT_HITS {
                         tracing::info!(
                             agent_id = %definition.id,
                             toolkit = %tk,
@@ -1007,7 +900,7 @@ async fn run_inner_loop(
         // `ToolCall` per parsed block so the rest of the loop can stay
         // uniform.
         let native_calls: Vec<ToolCall> = if force_text_mode {
-            let (_cleaned, parsed) = super::parse::parse_tool_calls(&response_text);
+            let (_cleaned, parsed) = super::super::parse::parse_tool_calls(&response_text);
             parsed
                 .into_iter()
                 .enumerate()
@@ -1056,7 +949,7 @@ async fn run_inner_loop(
             history.push(ChatMessage::assistant(response_text.clone()));
         } else {
             let assistant_history_content =
-                super::parse::build_native_assistant_history(&response_text, &native_calls);
+                super::super::parse::build_native_assistant_history(&response_text, &native_calls);
             history.push(ChatMessage::assistant(assistant_history_content));
         }
 
@@ -1209,931 +1102,10 @@ fn parse_tool_arguments(arguments: &str) -> serde_json::Value {
         .unwrap_or_else(|_| serde_json::Value::Object(Default::default()))
 }
 
-// ─────────────────────────────────────────────────────────────────────────────
-// Oversized-tool-result handoff (progressive disclosure)
-// ─────────────────────────────────────────────────────────────────────────────
-//
-// Typed sub-agents (integrations_agent in particular) regularly call tools that
-// return megabyte-scale payloads — `GMAIL_LIST_MESSAGES`, `NOTION_GET_PAGE`,
-// `GOOGLEDRIVE_LIST_FILES`. The default behaviour pushes that raw blob into
-// the sub-agent's history as a tool-result message, and the NEXT iteration
-// then ships the bloated history back to the provider where it hits the
-// model's context-length ceiling (observed: 276k-token prompt against a
-// 196 607-token window → backend 400).
-//
-// The summarizer dispatches to the model with the payload + an extraction
-// contract, but wiring it to fire on EVERY oversized result has two costs:
-// (1) we pay for a summarizer turn even when the sub-agent could answer
-// from a preview, and (2) aggressive compression sometimes drops the exact
-// identifier the agent needs for a follow-up call.
-//
-// Progressive disclosure fixes both: when a tool returns too much data,
-// we stash the full payload in a session-scoped cache, replace it in
-// history with a short placeholder (size + preview + result_id + how to
-// query it), and expose an `extract_from_result` tool that the sub-agent
-// can call with a targeted query. The summarizer only runs when the
-// sub-agent actually asks for a narrower view.
-
-/// Token threshold above which a tool result is routed to the handoff
-/// cache instead of being pushed into history raw. Token count is
-/// estimated at ~4 chars/token (mirrors
-/// [`crate::openhuman::agent::harness::payload_summarizer`] and
-/// [`crate::openhuman::tree_summarizer::types::estimate_tokens`]).
-///
-/// Raised from the original `20_000` to `50_000` so the clean Gmail /
-/// Notion envelopes emitted by provider post-processing fit through
-/// unchanged for normal workloads — only genuinely oversized results
-/// (bulk fetches, raw thread dumps) are routed through the
-/// `extract_from_result` path. The payload summarizer upstream of this
-/// stays disabled; `extract_from_result` itself still dispatches the
-/// summarizer sub-agent on demand when the sub-agent explicitly asks
-/// for a narrower view.
-const HANDOFF_OVERSIZE_THRESHOLD_TOKENS: usize = 50_000;
-
-/// Characters of the raw payload to surface in the placeholder preview.
-/// Enough for the sub-agent to recognise the shape (JSON keys, first
-/// record) and often small enough to answer trivial questions without a
-/// follow-up `extract_from_result` call.
-const HANDOFF_PREVIEW_CHARS: usize = 1500;
-
-/// Maximum entries per session. Bounded to keep memory use predictable
-/// on long-running sub-agents that might call many large tools. When
-/// over capacity we evict the oldest entry (FIFO); callers see "no
-/// cached result" for evicted ids and can either re-run the tool or
-/// ask the user/orchestrator to narrow the request.
-const HANDOFF_MAX_ENTRIES: usize = 8;
-
-/// Per-spawn cache of oversized tool payloads. One instance is built
-/// at the top of [`run_typed_mode`] and shared (via `Arc`) with both
-/// the inner tool-call loop (writes) and the `extract_from_result`
-/// tool (reads).
-#[derive(Default)]
-struct ResultHandoffCache {
-    inner: StdMutex<HandoffInner>,
-}
-
-#[derive(Default)]
-struct HandoffInner {
-    /// FIFO of inserted ids, used for eviction.
-    order: Vec<String>,
-    /// Content by id.
-    entries: HashMap<String, CachedResult>,
-    /// Monotonic counter for id generation within the session.
-    next_id: u64,
-}
-
-struct CachedResult {
-    tool_name: String,
-    content: String,
-}
-
-impl ResultHandoffCache {
-    fn new() -> Self {
-        Self::default()
-    }
-
-    /// Stash a payload and return a stable, short, grep-friendly id.
-    fn store(&self, tool_name: String, content: String) -> String {
-        let mut g = match self.inner.lock() {
-            Ok(g) => g,
-            Err(poisoned) => poisoned.into_inner(),
-        };
-        g.next_id = g.next_id.saturating_add(1);
-        let id = format!("res_{:x}", g.next_id);
-        g.order.push(id.clone());
-        g.entries
-            .insert(id.clone(), CachedResult { tool_name, content });
-        while g.order.len() > HANDOFF_MAX_ENTRIES {
-            let evicted = g.order.remove(0);
-            g.entries.remove(&evicted);
-        }
-        id
-    }
-
-    fn get(&self, result_id: &str) -> Option<CachedResult> {
-        let g = self.inner.lock().ok()?;
-        g.entries.get(result_id).map(|r| CachedResult {
-            tool_name: r.tool_name.clone(),
-            content: r.content.clone(),
-        })
-    }
-}
-
-/// Build the placeholder text that replaces an oversized tool result in
-/// the sub-agent's history. Shows the payload size (estimated tokens
-/// and raw bytes), a preview, and a call shape for the
-/// `extract_from_result` tool. The sub-agent decides whether to answer
-/// from the preview or dispatch the extractor.
-///
-/// Token count is estimated at ~4 chars/token (same heuristic as the
-/// trigger threshold in `HANDOFF_OVERSIZE_THRESHOLD_TOKENS`), so the
-/// unit the sub-agent sees matches the unit the runtime used to decide
-/// to hand off in the first place.
-fn build_handoff_placeholder(tool_name: &str, result_id: &str, raw: &str) -> String {
-    let preview: String = raw.chars().take(HANDOFF_PREVIEW_CHARS).collect();
-    let raw_tokens = raw.len().div_ceil(4);
-    format!(
-        "[oversized tool output: {raw_tokens} tokens ({raw_bytes} bytes) — stashed as result_id=\"{result_id}\"]\n\
-         Preview (first {preview_chars} chars):\n{preview}\n\n\
-         If the preview does not answer your task, call:\n\
-         extract_from_result(result_id=\"{result_id}\", query=\"<specific question>\")\n\
-         Good queries name the exact fields/identifiers you need \
-         (e.g. \"subject and sender of the 5 most recent messages\"). \
-         Tool: {tool_name}",
-        raw_bytes = raw.len(),
-        preview_chars = preview.chars().count(),
-    )
-}
-
-/// Sub-agent-side tool that answers a targeted query against a payload
-/// previously stashed via [`build_handoff_placeholder`]. Calls the
-/// configured provider directly — no sub-agent spawn — and returns the
-/// extracted answer inline.
-///
-/// We deliberately do **not** dispatch the `summarizer` archetype here:
-/// an agent-level wrapper brings system-prompt scaffolding, tool-loop
-/// plumbing, and an extra inference round that this job doesn't need.
-/// A single `chat_with_system` call is sufficient for the
-/// "extract facts matching a query" workload, gets the same provider
-/// routing (via the `"summarization-v1"` model id) for free, and
-/// removes the last place where the handoff path could trigger
-/// recursive sub-agent dispatch.
-struct ExtractFromResultTool {
-    cache: Arc<ResultHandoffCache>,
-    provider: Arc<dyn Provider>,
-    /// Workspace root for transcript writes.
-    workspace_dir: std::path::PathBuf,
-    /// Parent session chain joined with `__`, e.g.
-    /// `"1700000000_orchestrator__1700000005_1234_integrations_agent_abc"`.
-    /// Extract-call transcripts append a unique per-call suffix to this.
-    parent_chain: String,
-    /// Logical agent id that owns the calls (e.g. `"integrations_agent"`).
-    /// Only used to compose a descriptive `agent_name` in transcript meta.
-    owner_agent_id: String,
-    /// Monotonic counter so repeated calls within the same millisecond
-    /// still land on distinct transcript files.
-    call_seq: StdMutex<u64>,
-}
-
-impl ExtractFromResultTool {
-    fn new(
-        cache: Arc<ResultHandoffCache>,
-        provider: Arc<dyn Provider>,
-        workspace_dir: std::path::PathBuf,
-        parent_chain: String,
-        owner_agent_id: String,
-    ) -> Self {
-        Self {
-            cache,
-            provider,
-            workspace_dir,
-            parent_chain,
-            owner_agent_id,
-            call_seq: StdMutex::new(0),
-        }
-    }
-
-    fn next_call_seq(&self) -> u64 {
-        let mut guard = self
-            .call_seq
-            .lock()
-            .expect("extract_from_result call_seq mutex poisoned");
-        *guard = guard.saturating_add(1);
-        *guard
-    }
-}
-
-/// Model id used for `extract_from_result` LLM calls. Mirrors the
-/// resolution `ModelSpec::Hint("summarization").resolve(...)` would have
-/// produced for the retired summarizer sub-agent so routing table
-/// entries that targeted the summarizer continue to apply.
-const EXTRACT_MODEL_ID: &str = "summarization-v1";
-
-/// Temperature for extraction calls. Low but non-zero so the model can
-/// pick reasonable phrasings when rewriting identifiers into a compact
-/// answer, without straying into creative territory.
-const EXTRACT_TEMPERATURE: f64 = 0.2;
-
-/// System prompt fed to the provider on every `extract_from_result`
-/// call. Lifted in spirit from the old `summarizer` agent's prompt but
-/// trimmed to the core extraction contract — no fluff about iteration
-/// budgets or sub-agent roles because this is a pure tool call.
-const EXTRACT_SYSTEM_PROMPT: &str = "\
-You are an extraction assistant. A larger tool output is provided below. \
-Return ONLY the specific facts the user's query asks for. \
-Preserve identifiers verbatim (ids, urls, emails, timestamps, prices). \
-Be compact: no preamble, no commentary, no apologies, no meta-statements. \
-If the payload contains nothing relevant to the query, reply with an \
-empty string — do not invent information.";
-
-/// Persist a single extract-from-result LLM round-trip as its own
-/// transcript file under `session_raw/DDMMYYYY/{stem}.jsonl` (+ `.md`).
-///
-/// Best-effort: transcript failures are logged and swallowed so a
-/// readable-log hiccup never blocks the extraction itself. Appends a
-/// short suffix to the parent chain so every call lands on a distinct
-/// file (sibling extract calls within the same tool invocation still
-/// get unique stems).
-fn write_extract_transcript(
-    workspace_dir: &std::path::Path,
-    parent_chain: &str,
-    owner_agent_id: &str,
-    call_seq: u64,
-    chunk_label: Option<&str>,
-    system_prompt: &str,
-    user_prompt: &str,
-    assistant_output: Result<&str, &str>,
-    model: &str,
-) {
-    use super::session::transcript::{
-        resolve_keyed_transcript_path, write_transcript, MessageUsage, TranscriptMeta, TurnUsage,
-    };
-    use crate::openhuman::providers::ChatMessage;
-
-    let now = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default();
-    let unix_ts = now.as_secs();
-    let nanos = now.subsec_nanos();
-    let chunk_tag = match chunk_label {
-        Some(label) => format!("_{label}"),
-        None => String::new(),
-    };
-    let stem = format!(
-        "{parent_chain}__extract_{unix_ts}_{nanos:09}_{call_seq:04}{chunk_tag}"
-    );
-
-    let path = match resolve_keyed_transcript_path(workspace_dir, &stem) {
-        Ok(p) => p,
-        Err(e) => {
-            tracing::warn!(
-                error = %e,
-                stem = %stem,
-                "[extract_from_result] could not resolve transcript path; skipping transcript"
-            );
-            return;
-        }
-    };
-
-    let (assistant_text, is_error) = match assistant_output {
-        Ok(text) => (text.to_string(), false),
-        Err(err) => (format!("[error] {err}"), true),
-    };
-
-    let messages = vec![
-        ChatMessage {
-            role: "system".into(),
-            content: system_prompt.to_string(),
-        },
-        ChatMessage {
-            role: "user".into(),
-            content: user_prompt.to_string(),
-        },
-        ChatMessage {
-            role: "assistant".into(),
-            content: assistant_text,
-        },
-    ];
-
-    // Token counts aren't surfaced by `chat_with_system`; leave cost /
-    // usage fields zeroed and let the backend's own telemetry fill in
-    // the blanks when we wire richer accounting later.
-    let ts_rfc3339 = chrono::Utc::now().to_rfc3339();
-    let turn_usage = TurnUsage {
-        model: model.to_string(),
-        usage: MessageUsage {
-            input: 0,
-            output: 0,
-            cached_input: 0,
-            cost_usd: 0.0,
-        },
-        ts: ts_rfc3339.clone(),
-    };
-
-    let meta = TranscriptMeta {
-        agent_name: format!("{owner_agent_id}::extract_from_result"),
-        dispatcher: "native".into(),
-        created: ts_rfc3339.clone(),
-        updated: ts_rfc3339,
-        turn_count: 1,
-        input_tokens: 0,
-        output_tokens: 0,
-        cached_input_tokens: 0,
-        charged_amount_usd: 0.0,
-    };
-
-    if let Err(e) = write_transcript(&path, &messages, &meta, Some(&turn_usage)) {
-        tracing::warn!(
-            error = %e,
-            path = %path.display(),
-            "[extract_from_result] transcript write failed"
-        );
-    } else {
-        tracing::debug!(
-            path = %path.display(),
-            is_error,
-            "[extract_from_result] transcript written"
-        );
-    }
-}
-
-#[async_trait]
-impl Tool for ExtractFromResultTool {
-    fn name(&self) -> &str {
-        "extract_from_result"
-    }
-
-    fn description(&self) -> &str {
-        "Answer a targeted question against an oversized tool output that was \
-         stashed under a `result_id` handle. Use this when a previous tool call \
-         returned a placeholder like `result_id=\"res_1\"` because its raw output \
-         was too large to show inline. Pass the handle plus a natural-language \
-         `query` naming the exact facts/identifiers you need; returns only the \
-         extracted answer, not the full payload. Multiple queries against the \
-         same `result_id` are allowed — each one is independent."
-    }
-
-    fn parameters_schema(&self) -> serde_json::Value {
-        serde_json::json!({
-            "type": "object",
-            "properties": {
-                "result_id": {
-                    "type": "string",
-                    "description": "The handle emitted in the oversized tool output placeholder (e.g. `res_1`)."
-                },
-                "query": {
-                    "type": "string",
-                    "description": "Natural-language question naming the exact facts or identifiers to extract. Be specific."
-                }
-            },
-            "required": ["result_id", "query"]
-        })
-    }
-
-    fn category(&self) -> ToolCategory {
-        ToolCategory::System
-    }
-
-    async fn execute(&self, args: serde_json::Value) -> anyhow::Result<ToolResult> {
-        let result_id = args.get("result_id").and_then(|v| v.as_str()).unwrap_or("");
-        let query = args.get("query").and_then(|v| v.as_str()).unwrap_or("");
-
-        if result_id.is_empty() || query.is_empty() {
-            return Ok(ToolResult::error(
-                "extract_from_result requires non-empty `result_id` and `query`.",
-            ));
-        }
-
-        let cached = match self.cache.get(result_id) {
-            Some(c) => c,
-            None => {
-                return Ok(ToolResult::error(format!(
-                    "No cached result found for id '{result_id}'. The handle may have been evicted (cache holds the {HANDOFF_MAX_ENTRIES} most recent entries). Re-run the original tool to get a fresh handle."
-                )));
-            }
-        };
-
-        // Fast path: payload fits in a single provider turn.
-        if cached.content.len() <= SUMMARIZER_CHUNK_CHAR_BUDGET {
-            tracing::debug!(
-                tool = %cached.tool_name,
-                bytes = cached.content.len(),
-                "[extract_from_result] single-shot extraction"
-            );
-            return self
-                .extract_single_shot(&cached.tool_name, &cached.content, query)
-                .await;
-        }
-
-        // Slow path: chunk + parallel map. A single call on a payload
-        // large enough to need the handoff (hundreds of KB common for
-        // Gmail / Notion list operations) risks either (a) overflowing
-        // the extraction model's context window, or (b) a low-quality
-        // single-pass answer that misses facts near the tail. Splitting
-        // into budgeted chunks and running them in parallel keeps each
-        // call under its context budget and usually finishes faster
-        // than a sequential single-shot call on the whole blob.
-        //
-        // No reduce stage: per-chunk extracts are concatenated in
-        // original chunk order. A reduce LLM call adds latency (often
-        // the slowest single turn) and becomes a single point of
-        // failure when the upstream provider stalls. For
-        // listing/extraction queries concatenation is equivalent; for
-        // top-N / global-ordering queries the caller can post-process.
-        let chunks = chunk_content(&cached.content, SUMMARIZER_CHUNK_CHAR_BUDGET);
-        tracing::info!(
-            tool = %cached.tool_name,
-            total_bytes = cached.content.len(),
-            chunk_count = chunks.len(),
-            chunk_budget = SUMMARIZER_CHUNK_CHAR_BUDGET,
-            "[extract_from_result] chunked extraction"
-        );
-
-        // Map stage: each chunk extracts items matching `query` from
-        // ITS OWN slice only. Dispatched with bounded concurrency —
-        // `buffer_unordered(MAP_CONCURRENCY)` keeps at most N calls in
-        // flight at any time. Fully parallel `join_all` was generating
-        // 504-gateway-timeout storms from the staging proxy when 7+
-        // concurrent calls piled onto the upstream; batching at 3
-        // trades some wall-clock time for reliability.
-        const MAP_CONCURRENCY: usize = 3;
-        let total_chunks = chunks.len();
-
-        // Each chunk gets its own monotonic call_seq so sibling
-        // transcripts written in parallel still land on distinct files.
-        let call_seq_base = self.next_call_seq();
-        let workspace_dir = self.workspace_dir.clone();
-        let parent_chain = self.parent_chain.clone();
-        let owner_agent_id = self.owner_agent_id.clone();
-
-        // Consume `chunks` with `into_iter` so each async block owns
-        // its `String` — `buffer_unordered` polls the stream lazily
-        // and needs futures with no borrows into the enclosing scope.
-        let map_futures = chunks.into_iter().enumerate().map(|(i, chunk)| {
-            let provider = self.provider.clone();
-            let tool_name = cached.tool_name.clone();
-            let query = query.to_string();
-            let workspace_dir = workspace_dir.clone();
-            let parent_chain = parent_chain.clone();
-            let owner_agent_id = owner_agent_id.clone();
-            async move {
-                let user_prompt = format!(
-                    "Tool name: {tool_name}\nChunk {idx} of {total}\n\n\
-                     Query: {query}\n\n\
-                     This is one slice of a larger tool output. Extract ONLY \
-                     items in THIS slice that match the query. Preserve \
-                     identifiers verbatim. Return an empty string if nothing \
-                     in this slice is relevant.\n\n\
-                     --- BEGIN SLICE ---\n{chunk}\n--- END SLICE ---",
-                    idx = i + 1,
-                    total = total_chunks,
-                );
-                let result = provider
-                    .chat_with_system(
-                        Some(EXTRACT_SYSTEM_PROMPT),
-                        &user_prompt,
-                        EXTRACT_MODEL_ID,
-                        EXTRACT_TEMPERATURE,
-                    )
-                    .await;
-
-                // Persist this chunk's transcript before returning, so
-                // a partial failure higher up the stream still leaves
-                // an auditable record on disk.
-                let transcript_input: Result<&str, String> = match &result {
-                    Ok(text) => Ok(text.as_str()),
-                    Err(e) => Err(e.to_string()),
-                };
-                let chunk_label = format!("chunk{:03}of{:03}", i + 1, total_chunks);
-                write_extract_transcript(
-                    &workspace_dir,
-                    &parent_chain,
-                    &owner_agent_id,
-                    call_seq_base,
-                    Some(&chunk_label),
-                    EXTRACT_SYSTEM_PROMPT,
-                    &user_prompt,
-                    match &transcript_input {
-                        Ok(s) => Ok(*s),
-                        Err(s) => Err(s.as_str()),
-                    },
-                    EXTRACT_MODEL_ID,
-                );
-
-                (i, result)
-            }
-        });
-
-        let mut map_results: Vec<(usize, _)> = futures::stream::iter(map_futures)
-            .buffer_unordered(MAP_CONCURRENCY)
-            .collect()
-            .await;
-        // `buffer_unordered` yields futures in completion order; restore
-        // original chunk order so the concatenated output matches the
-        // natural ordering of the underlying tool result (e.g. Notion's
-        // reverse-chrono page list).
-        map_results.sort_by_key(|(i, _)| *i);
-
-        let partials: Vec<String> = map_results
-            .into_iter()
-            .filter_map(|(i, r)| match r {
-                Ok(text) => {
-                    let trimmed = text.trim();
-                    if trimmed.is_empty() {
-                        None
-                    } else {
-                        Some(trimmed.to_string())
-                    }
-                }
-                Err(e) => {
-                    tracing::warn!(
-                        chunk_idx = i,
-                        error = %e,
-                        "[extract_from_result] map-stage provider call failed; dropping partial"
-                    );
-                    None
-                }
-            })
-            .collect();
-
-        if partials.is_empty() {
-            return Ok(ToolResult::error(
-                "extract_from_result: no matching content found across any chunk",
-            ));
-        }
-
-        // Concatenate per-chunk summaries in original chunk order.
-        // `join` with a single partial yields it unchanged (no trailing
-        // separator), so no special-case is needed.
-        Ok(ToolResult::success(partials.join("\n\n---\n\n")))
-    }
-}
-
-impl ExtractFromResultTool {
-    async fn extract_single_shot(
-        &self,
-        tool_name: &str,
-        content: &str,
-        query: &str,
-    ) -> anyhow::Result<ToolResult> {
-        let user_prompt = format!(
-            "Tool name: {tool_name}\n\nQuery: {query}\n\n\
-             Raw tool output follows. Extract ONLY the information the query \
-             asks for.\n\n\
-             --- BEGIN ---\n{content}\n--- END ---",
-        );
-
-        let call_seq = self.next_call_seq();
-        let provider_result = self
-            .provider
-            .chat_with_system(
-                Some(EXTRACT_SYSTEM_PROMPT),
-                &user_prompt,
-                EXTRACT_MODEL_ID,
-                EXTRACT_TEMPERATURE,
-            )
-            .await;
-
-        // Persist the transcript before returning — the LLM call cost
-        // tokens regardless of whether we ultimately return success.
-        let transcript_input: Result<&str, String> = match &provider_result {
-            Ok(text) => Ok(text.as_str()),
-            Err(e) => Err(e.to_string()),
-        };
-        write_extract_transcript(
-            &self.workspace_dir,
-            &self.parent_chain,
-            &self.owner_agent_id,
-            call_seq,
-            None,
-            EXTRACT_SYSTEM_PROMPT,
-            &user_prompt,
-            match &transcript_input {
-                Ok(s) => Ok(*s),
-                Err(s) => Err(s.as_str()),
-            },
-            EXTRACT_MODEL_ID,
-        );
-
-        match provider_result {
-            Ok(text) => {
-                let trimmed = text.trim();
-                if trimmed.is_empty() {
-                    Ok(ToolResult::error(
-                        "extract_from_result: provider returned an empty response",
-                    ))
-                } else {
-                    Ok(ToolResult::success(trimmed.to_string()))
-                }
-            }
-            Err(e) => Ok(ToolResult::error(format!(
-                "extract_from_result: provider call failed: {e}"
-            ))),
-        }
-    }
-}
-
-/// Char budget per summarizer call. Chosen so a single chunk + prompt
-/// scaffolding + output stays well below the summarization model's
-/// context window (~196k tokens) — at ~4 chars/token that leaves
-/// comfortable headroom for the extraction contract and response.
-const SUMMARIZER_CHUNK_CHAR_BUDGET: usize = 60_000;
-
-/// Split `content` into chunks no larger than `budget` bytes, breaking
-/// at natural boundaries (blank lines, then single newlines) so the
-/// summarizer rarely sees a structure torn mid-record. Falls back to
-/// char-safe slicing for pathological single-line inputs.
-/// Strip common noise from tool outputs before they're stashed or chunked.
-///
-/// Agent tools frequently return raw HTML email bodies, inline SVG, base64
-/// data URIs, CSS/JS blocks, and collapsed whitespace — all of which bloat
-/// the handoff cache and waste summarizer context on tokens that carry
-/// zero semantic value for most extraction queries. Cleaning before the
-/// oversize check means (a) some payloads drop below threshold entirely
-/// and skip the extract pipeline, (b) chunked payloads fit more real
-/// content per chunk, and (c) summarizers see clean text instead of
-/// parsing around markup.
-///
-/// Applied generically to every tool output — no per-tool gating. The
-/// patterns below target universally-noisy markup; non-HTML outputs
-/// (plain JSON, plain text) are left essentially untouched because none
-/// of these regexes match.
-fn clean_tool_output(content: &str) -> String {
-    // Block-level containers with entirely useless payloads — match lazily
-    // across lines, case-insensitive. `(?s)` enables `.` matching `\n`.
-    static SCRIPT_RE: LazyLock<Regex> =
-        LazyLock::new(|| Regex::new(r"(?is)<script\b[^>]*>.*?</script\s*>").unwrap());
-    static STYLE_RE: LazyLock<Regex> =
-        LazyLock::new(|| Regex::new(r"(?is)<style\b[^>]*>.*?</style\s*>").unwrap());
-    static SVG_RE: LazyLock<Regex> =
-        LazyLock::new(|| Regex::new(r"(?is)<svg\b[^>]*>.*?</svg\s*>").unwrap());
-    static HTML_COMMENT_RE: LazyLock<Regex> =
-        LazyLock::new(|| Regex::new(r"(?s)<!--.*?-->").unwrap());
-    // `data:<mime>;base64,<...>` inline payloads — keep the agent aware
-    // an image/asset was there, but drop the bytes.
-    static DATA_URI_RE: LazyLock<Regex> =
-        LazyLock::new(|| Regex::new(r"(?i)data:[a-z0-9.+\-/]+;base64,[A-Za-z0-9+/=]+").unwrap());
-    // Everything remaining that still looks like an HTML tag — strip the
-    // tag but keep the text content. Deliberately greedy across attributes
-    // but not across `>` so we don't eat inter-tag content.
-    static HTML_TAG_RE: LazyLock<Regex> = LazyLock::new(|| Regex::new(r"<[^>]+>").unwrap());
-    // Runs of whitespace → single space; collapse vertical bloat.
-    static WS_RE: LazyLock<Regex> = LazyLock::new(|| Regex::new(r"[ \t\f\v]+").unwrap());
-    static BLANK_LINE_RE: LazyLock<Regex> = LazyLock::new(|| Regex::new(r"\n{3,}").unwrap());
-
-    let cleaned = SCRIPT_RE.replace_all(content, "");
-    let cleaned = STYLE_RE.replace_all(&cleaned, "");
-    let cleaned = SVG_RE.replace_all(&cleaned, "[svg]");
-    let cleaned = HTML_COMMENT_RE.replace_all(&cleaned, "");
-    let cleaned = DATA_URI_RE.replace_all(&cleaned, "[data-uri]");
-    let cleaned = HTML_TAG_RE.replace_all(&cleaned, "");
-    let cleaned = WS_RE.replace_all(&cleaned, " ");
-    let cleaned = BLANK_LINE_RE.replace_all(&cleaned, "\n\n");
-    cleaned.trim().to_string()
-}
-
-fn chunk_content(content: &str, budget: usize) -> Vec<String> {
-    if content.len() <= budget {
-        return vec![content.to_string()];
-    }
-
-    let mut chunks: Vec<String> = Vec::new();
-    let mut current = String::with_capacity(budget.min(content.len()));
-
-    let flush = |current: &mut String, chunks: &mut Vec<String>| {
-        if !current.is_empty() {
-            chunks.push(std::mem::take(current));
-        }
-    };
-
-    for line in content.lines() {
-        let projected = current.len() + line.len() + 1;
-        if projected > budget && !current.is_empty() {
-            flush(&mut current, &mut chunks);
-        }
-        if line.len() > budget {
-            // Single line exceeds budget (e.g. JSON with no formatting).
-            // Emit any pending content, then slice the line at char
-            // boundaries so we don't panic on multi-byte chars.
-            flush(&mut current, &mut chunks);
-            let mut remaining = line;
-            while !remaining.is_empty() {
-                let mut cut = budget.min(remaining.len());
-                while cut > 0 && !remaining.is_char_boundary(cut) {
-                    cut -= 1;
-                }
-                if cut == 0 {
-                    // Degenerate case — shouldn't happen for normal
-                    // text. Take the entire remaining line to avoid
-                    // an infinite loop.
-                    chunks.push(remaining.to_string());
-                    break;
-                }
-                chunks.push(remaining[..cut].to_string());
-                remaining = &remaining[cut..];
-            }
-        } else {
-            current.push_str(line);
-            current.push('\n');
-        }
-    }
-    flush(&mut current, &mut chunks);
-
-    chunks
-}
-
-/// Tight top-K ceiling for toolkits whose per-action JSON schemas are
-/// dense enough to blow through either Fireworks' 65 535-rule grammar
-/// cap (native mode) or the 196 607-token context cap (text mode) even
-/// before any tool results land in history. Determined empirically from
-/// the fixture dumps under `tests/fixtures/composio_*.json` and real
-/// staging failures — see the trace where Gmail at top-K=25 produced
-/// a 276k-token iter-1 prompt.
-const HEAVY_SCHEMA_TOOLKITS: &[&str] = &[
-    "gmail",
-    "notion",
-    "github",
-    "salesforce",
-    "hubspot",
-    "googledrive",
-    "googlesheets",
-    "googledocs",
-    "microsoftteams",
-];
-
-const TOOL_FILTER_TOP_K_DEFAULT: usize = 25;
-const TOOL_FILTER_TOP_K_HEAVY: usize = 12;
-
-/// Pick a top-K budget for the fuzzy filter based on how dense the
-/// toolkit's action schemas tend to be. Match is case-insensitive so
-/// we don't care whether the caller passed `"Gmail"` or `"gmail"`.
-fn top_k_for_toolkit(toolkit: &str) -> usize {
-    if HEAVY_SCHEMA_TOOLKITS
-        .iter()
-        .any(|t| t.eq_ignore_ascii_case(toolkit))
-    {
-        TOOL_FILTER_TOP_K_HEAVY
-    } else {
-        TOOL_FILTER_TOP_K_DEFAULT
-    }
-}
-
-/// Format a set of `ToolSpec`s as an XML tool-use protocol block
-/// appended to the system prompt in text mode. Mirrors
-/// [`crate::openhuman::agent::dispatcher::XmlToolDispatcher::prompt_instructions`]
-/// — same `<tool_call>{…}</tool_call>` format so the existing
-/// `parse_tool_calls` helper understands what the model emits.
-///
-/// Per-parameter rendering is intentionally **compact**: name, type, a
-/// "required" marker, and a short one-line description if present. We
-/// do **not** serialise the full JSON schema. Composio/Fireworks action
-/// schemas for toolkits like Gmail or Notion run multiple KB each —
-/// embedding them verbatim blows up the prompt past the model's
-/// context window (282k+ tokens for 26 Gmail tools vs a 196k cap).
-/// The compact listing keeps the model informed enough to call tools
-/// correctly while staying within budget. If the model needs deeper
-/// schema detail it can surface the error and the orchestrator will
-/// clarify on the next turn.
-pub(crate) fn build_text_mode_tool_instructions(_specs: &[ToolSpec]) -> String {
-    // The tool catalog is already rendered in the prompt's `## Tools`
-    // section (see `prompts::ToolsSection::build`) with full
-    // `Call as: NAME[arg|arg]` signatures. We previously also emitted
-    // an `### Available Tools` subsection here with a different
-    // formatting (`Parameters: name:type, ...`), which doubled the
-    // tool list bytes for text-mode agents — especially expensive for
-    // the integrations_agent toolkit-scoped spawns (~50 actions ×
-    // 2 listings). Keep only the protocol explanation; the tool
-    // catalog itself comes from the prompt template.
-    let mut out = String::new();
-    out.push_str("## Tool Use Protocol\n\n");
-    out.push_str(
-        "To use a tool, wrap a JSON object in <tool_call></tool_call> tags. \
-         Do not nest tags. Emit one tag per call; you can emit multiple tags \
-         in the same response if you need to run calls in parallel.\n\n",
-    );
-    out.push_str(
-        "```\n<tool_call>\n{\"name\": \"tool_name\", \"arguments\": {\"param\": \"value\"}}\n</tool_call>\n```\n",
-    );
-    out
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Tool filtering
-// ─────────────────────────────────────────────────────────────────────────────
-
-/// Returns indices into `parent_tools` for the tools the sub-agent may
-/// invoke. Index-based filtering avoids cloning `Box<dyn Tool>` (which
-/// isn't Clone) and lets us reuse the parent's existing instances.
-///
-/// Filters are applied in this order (shorter-circuit first):
-/// 1. `disallowed` — explicit deny list.
-/// 2. `skill_filter` — restrict to tools named `{skill}__*`.
-/// 3. `scope` — `Wildcard` (everything remaining) or `Named` allowlist.
-///
-/// Exposed `pub(crate)` so the debug dump path in
-/// [`crate::openhuman::context::debug_dump`] shares the exact same
-/// filter logic as the live runner — previously debug_dump carried a
-/// "standalone copy" which drifted over time.
-/// Tools that must never be visible to any agent except `welcome`.
-///
-/// `complete_onboarding` flips the onboarding-complete flag in
-/// workspace config and is the terminal step of the welcome flow;
-/// every other agent must route the user back to the welcome agent
-/// rather than call it directly. Central list here so both the main
-/// agent builder ([`crate::openhuman::agent::harness::session::builder`])
-/// and the subagent runner apply the same guard.
-pub(crate) fn is_welcome_only_tool(name: &str) -> bool {
-    matches!(name, "complete_onboarding")
-}
-
-pub(crate) fn filter_tool_indices(
-    parent_tools: &[Box<dyn Tool>],
-    scope: &ToolScope,
-    disallowed: &[String],
-    skill_filter: Option<&str>,
-) -> Vec<usize> {
-    let disallow_set: HashSet<&str> = disallowed.iter().map(|s| s.as_str()).collect();
-    let skill_prefix = skill_filter.map(|s| format!("{s}__"));
-
-    parent_tools
-        .iter()
-        .enumerate()
-        .filter(|(_, tool)| {
-            let name = tool.name();
-            if disallow_set.contains(name) {
-                return false;
-            }
-            if let Some(prefix) = skill_prefix.as_deref() {
-                if !name.starts_with(prefix) {
-                    return false;
-                }
-            }
-            match scope {
-                ToolScope::Wildcard => true,
-                ToolScope::Named(allowed) => allowed.iter().any(|n| n == name),
-            }
-        })
-        .map(|(i, _)| i)
-        .collect()
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Prompt loading + composition
-// ─────────────────────────────────────────────────────────────────────────────
-
-/// Resolve a [`PromptSource`] to its raw markdown body. Inline sources
-/// return immediately, `Dynamic` calls the builder with the supplied
-/// [`PromptContext`], `File` sources are read from disk relative to the
-/// workspace `prompts/` directory or the agent crate's bundled prompts.
-///
-/// Exposed `pub(crate)` so the debug dump path in
-/// [`crate::openhuman::context::debug_dump`] loads prompts through the
-/// exact same code the runner uses — no parallel body-loading logic.
-pub(crate) fn load_prompt_source(
-    source: &PromptSource,
-    ctx: &PromptContext<'_>,
-) -> Result<String, SubagentRunError> {
-    let workspace_dir = ctx.workspace_dir;
-    match source {
-        PromptSource::Inline(body) => Ok(body.clone()),
-        PromptSource::Dynamic(build) => build(ctx).map_err(|e| SubagentRunError::PromptLoad {
-            path: format!("<dynamic:{}>", ctx.agent_id),
-            source: std::io::Error::new(std::io::ErrorKind::Other, e.to_string()),
-        }),
-        PromptSource::File { path } => {
-            // Try the workspace's `agent/prompts/` first (so users can
-            // override built-in prompts), then fall back to the crate's
-            // own bundled prompts via `include_str!`-style lookup.
-            let workspace_path = workspace_dir.join("agent").join("prompts").join(path);
-            if workspace_path.is_file() {
-                return std::fs::read_to_string(&workspace_path).map_err(|e| {
-                    SubagentRunError::PromptLoad {
-                        path: workspace_path.display().to_string(),
-                        source: e,
-                    }
-                });
-            }
-            // Built-in prompt fallback. The agent prompts directory is
-            // already shipped at `src/openhuman/agent/prompts/` and
-            // included in the binary via the `IdentitySection` workspace
-            // file write — so we re-use that scaffolding by reading from
-            // `<workspace>/<filename>` after the parent agent has
-            // bootstrapped its workspace files. For sub-agent
-            // archetype prompts (e.g. `archetypes/researcher.md`),
-            // we look up by basename in the workspace, then accept
-            // missing files as an empty body (the runner will fall
-            // back to a generic role hint).
-            let workspace_root_path = workspace_dir.join(path);
-            if workspace_root_path.is_file() {
-                return std::fs::read_to_string(&workspace_root_path).map_err(|e| {
-                    SubagentRunError::PromptLoad {
-                        path: workspace_root_path.display().to_string(),
-                        source: e,
-                    }
-                });
-            }
-            tracing::warn!(
-                path = %path,
-                "[subagent_runner] archetype prompt file not found, using empty body"
-            );
-            Ok(String::new())
-        }
-    }
-}
-
-// Note: the narrow sub-agent prompt renderer lives in
-// `crate::openhuman::context::prompt::render_subagent_system_prompt`
-// so every system-prompt-building call-site — main agents, sub-agents,
-// channel runtimes — shares one module.
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Tests
-// ─────────────────────────────────────────────────────────────────────────────
-
 #[cfg(test)]
 mod tests {
-    use super::super::definition::ModelSpec;
     use super::*;
+    use crate::openhuman::agent::harness::definition::{ModelSpec, ToolScope};
 
     fn make_def_named_tools(names: &[&str]) -> AgentDefinition {
         AgentDefinition {
@@ -2155,12 +1127,12 @@ mod tests {
             extra_tools: vec![],
             max_iterations: 5,
             timeout_secs: None,
-            sandbox_mode: super::super::definition::SandboxMode::None,
+            sandbox_mode: crate::openhuman::agent::harness::definition::SandboxMode::None,
             background: false,
             uses_fork_context: false,
             subagents: vec![],
             delegate_name: None,
-            source: super::super::definition::DefinitionSource::Builtin,
+            source: crate::openhuman::agent::harness::definition::DefinitionSource::Builtin,
         }
     }
 
@@ -2253,7 +1225,9 @@ mod tests {
 
     // ── End-to-end runner tests with mock provider ────────────────────────
 
-    use super::super::fork_context::{with_fork_context, with_parent_context};
+    use crate::openhuman::agent::harness::fork_context::{
+        with_fork_context, with_parent_context,
+    };
     use crate::openhuman::providers::{
         ChatRequest as PChatRequest, ChatResponse, Provider, ToolCall,
     };
@@ -2655,7 +1629,7 @@ mod tests {
             fork_task_prompt: "ANALYSE THIS BRANCH".into(),
         };
 
-        let def = super::super::builtin_definitions::fork_definition();
+        let def = crate::openhuman::agent::harness::builtin_definitions::fork_definition();
 
         let outcome = with_parent_context(parent, async move {
             with_fork_context(fork, async {
@@ -2694,7 +1668,7 @@ mod tests {
     async fn fork_mode_errors_when_no_fork_context() {
         let provider = ScriptedProvider::new(vec![text_response("unused")]);
         let parent = make_parent(provider, vec![stub("file_read")]);
-        let def = super::super::builtin_definitions::fork_definition();
+        let def = crate::openhuman::agent::harness::builtin_definitions::fork_definition();
 
         let result = with_parent_context(parent, async {
             run_subagent(&def, "x", SubagentRunOptions::default()).await
